task number 350

+++ Ch1pchilinka/26. Remove Duplicates from Sorted Array /main.cpp
@@ -5,8 +5,8 @@ class Solution {
 public:
   int removeDuplicates(std::vector<int>& nums) {
     int ind = 1;
-    int n = nums.size();
-    for (int i = 0; i < n - 1; ++i) {
+    size_t n = nums.size();
+    for (size_t i = 0; i < n - 1; ++i) {
       if (nums[i] != nums[i + 1]) {
         nums[ind++] = nums[i + 1];
       }
@@ -16,16 +16,16 @@ public:
 };
 
 int main() {
-  int n;
+  size_t n;
   std::cin >> n;
   std::vector<int> arr(n);
-  for (int i = 0; i < n; ++i) {
+  for (size_t i = 0; i < n; ++i) {
     std::cin >> arr[i];
   }
   Solution s;
   std::cout << s.removeDuplicates(arr) << "\n";
 
-  for (int i = 0; i < n; ++i) {
+  for (size_t i = 0; i < n; ++i) {
     std::cout << arr[i] << " ";
   }
   return 0;
